+++ core/sr-api-macros/tests/ui/missing_path_for_trait.rs
@@ -1,0 +1,26 @@
+use runtime_primitives::traits::GetNodeBlockType;
+use test_client::runtime::Block;
+use client::{decl_runtime_apis, impl_runtime_apis};
+
+/// The declaration of the `Runtime` type and the implementation of the `GetNodeBlockType`
+/// trait are done by the `construct_runtime!` macro in a real runtime.
+struct Runtime {}
+impl GetNodeBlockType for Runtime {
+	type NodeBlock = Block;
+}
+
+decl_runtime_apis! {
+	pub trait Api {
+		fn test(data: u64);
+	}
+}
+
+impl_runtime_apis! {
+	impl Api<Block> for Runtime {
+		fn test(data: u64) {
+			unimplemented!()
+		}
+	}
+}
+
+fn main() {}
