+++ core/sr-api-macros/tests/ui/missing_path_for_trait.stderr
@@ -1,0 +1,5 @@
+error: The implemented trait has to be referenced with a path, e.g. `impl client::Core for Runtime`.
+  --> $DIR/missing_path_for_trait.rs:19:7
+   |
+19 |     impl Api<Block> for Runtime {
+   |          ^^^
